Create abstract Recap class
