+++ spec/models/requests/service_eligibility/recap/in_library_spec.rb
@@ -2,6 +2,7 @@
 require 'rails_helper'
 
 RSpec.describe Requests::ServiceEligibility::Recap::InLibrary, requests: true do
+  let(:user) { FactoryBot.create(:user) }
   describe '#eligible?' do
     it 'returns true if all criteria are met' do
       requestable = instance_double(Requests::Requestable)
@@ -13,7 +14,7 @@ RSpec.describe Requests::ServiceEligibility::Recap::InLibrary, requests: true do
         recap_edd?: true,
         recap_pf?: true
       )
-      eligibility = described_class.new(requestable:)
+      eligibility = described_class.new(requestable:, user:)
 
       expect(eligibility.eligible?).to be(true)
     end
@@ -28,7 +29,7 @@ RSpec.describe Requests::ServiceEligibility::Recap::InLibrary, requests: true do
         recap_edd?: true,
         recap_pf?: false
       )
-      eligibility = described_class.new(requestable:)
+      eligibility = described_class.new(requestable:, user:)
 
       expect(eligibility.eligible?).to be(false)
     end
